применене FSD в проекте
