+++ public/index.html
@@ -12,7 +12,7 @@
   <title>React App</title>
 </head>
 
-<body>
+<body class="custom-scrollbar">
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
   <div id="modal-root"></div>
