+++ src/assets/styles/base.css
@@ -58,6 +58,7 @@ button {
   cursor: pointer;
   outline: none;
 }
+
 body {
   margin: 0;
   font-family: Roboto, -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu",
