+++ src/components/Card/Card.module.css
@@ -22,64 +22,6 @@
   height: 100%;
 }
 
-.info {
-  position: relative;
-  z-index: var(--z-index-s);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 12px;
-  gap: 8px;
-}
-.block {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: white;
-  border-radius: var(--border-radius-10);
-}
-
-.blockButton {
-  composes: block;
-  margin: auto;
-}
-
-.interaction {
-  display: flex;
-  gap: 5px;
-}
-
-.interaction .blockMessage {
-  position: relative;
-  visibility: hidden;
-  pointer-events: none;
-  opacity: 0;
-  left: -100%;
-  transition: var(--transition-duration-300) ease-in-out;
-}
-
-.icon {
-  background-color: white;
-  margin: 8px;
-  stroke: var(--color-simulator-gray-text);
-  transition: stroke var(--transition-duration-300) ease-in-out;
-}
-
-.iconActive {
-  fill: var(--color-main-blue);
-  stroke: var(--color-main-blue);
-}
-
-.tag {
-  font: var(--font-Roboto-14-normal);
-  padding: 8px 12px;
-}
-
-.message {
-  font: var(--font-Roboto-10-normal);
-  padding: 16px 8px;
-}
-
 .description {
   padding: 20px 24px 24px 24px;
   cursor: pointer;
@@ -118,28 +60,5 @@
   .card:hover {
     transform: translateY(-2px);
     box-shadow: 0 4px 6px -7px rgb(0 0 0 / 10%), 0 4px 30px -9px rgb(0 0 0 / 20%);
-  }
-
-  @media screen and (min-width: 767.5px) {
-    .interaction:hover .blockMessage {
-      visibility: visible;
-      pointer-events: auto;
-      opacity: 1;
-      left: 0;
-    }
-  }
-
-  .blockButton:hover .icon {
-    stroke: var(--color-main-primary);
-  }
-
-  .blockButton:hover .iconActive {
-    stroke: var(--color-main-blue);
-  }
-}
-
-@media screen and (max-width: 767px) {
-  .blockMessage {
-    display: none;
   }
 }
