+++ src/components/CardInfo/CardInfo.module.css
@@ -1,0 +1,82 @@
+.info {
+  position: relative;
+  z-index: var(--z-index-s);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 12px;
+  gap: 8px;
+}
+.block {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: white;
+  border-radius: var(--border-radius-10);
+  box-shadow: 0px 10px 22px rgba(59, 74, 116, 0.1);
+}
+
+.blockButton {
+  composes: block;
+  margin: auto;
+}
+.interaction {
+  display: flex;
+  gap: 5px;
+}
+
+.interaction .blockMessage {
+  position: relative;
+  visibility: hidden;
+  pointer-events: none;
+  opacity: 0;
+  left: -100%;
+  transition: var(--transition-duration-300) ease-in-out;
+}
+
+.icon {
+  background-color: white;
+  margin: 8px;
+  stroke: var(--color-simulator-gray-text);
+  transition: stroke var(--transition-duration-300) ease-in-out;
+}
+
+.iconActive {
+  fill: var(--color-main-blue);
+  stroke: var(--color-main-blue);
+}
+
+.tag {
+  font: var(--font-Roboto-14-normal);
+  padding: 8px 12px;
+}
+
+.message {
+  font: var(--font-Roboto-10-normal);
+  padding: 16px 8px;
+}
+
+@media (any-hover: hover) {
+  @media screen and (min-width: 767.5px) {
+    .interaction:hover .blockMessage {
+      visibility: visible;
+      pointer-events: auto;
+      opacity: 1;
+      left: 0;
+    }
+  }
+
+  .blockButton:hover .icon {
+    stroke: var(--color-main-primary);
+  }
+
+  .blockButton:hover .iconActive {
+    stroke: var(--color-main-blue);
+  }
+}
+
+@media screen and (max-width: 767px) {
+  .blockMessage {
+    display: none;
+  }
+}
