+++ src/entities/input/index.js
@@ -1,0 +1,2 @@
+export * from "./ui";
+export * as inputLib from "./lib";
