+++ src/entities/input/lib/utils.js
@@ -1,0 +1,10 @@
+export const errors = {
+  tooShortName: "Имя должно содержать больше 2-х символов",
+  tooLongName: "У нас больше нет символов ヽ(°□° )ノ",
+  tooShortPassword: "Надо бы понадежнее - введите мин. 8 символов",
+  tooLongPassword: "Не, ну это уже перебор",
+  patternMismatchName: "Ипользовать можно буквы и цифры, первый символ обязательно буква",
+  patternMismatchEmail: "Неверный формат Email",
+  patternMismatchPassword: "Доступны только строчные и прописные латинские буквы, цифры",
+  valueMissing: "Обязательное поле",
+};
